requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi @ file:///croot/certifi_1720453481653/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
charset-normalizer==3.1.0
click @ file:///croot/click_1698129812380/work
cloudpickle @ file:///croot/cloudpickle_1721657346512/work
cmake==3.26.0
coloredlogs==15.0.1
comm==0.1.2
conda==4.14.0
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1714660666131/work
cssselect2==0.7.0
cssutils==2.11.1
cycler==0.11.0
cytoolz @ file:///croot/cytoolz_1701723583781/work
dask @ file:///croot/dask-core_1683065217061/work
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
diffvg==0.0.1
easydict==1.13
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.0
fonttools==4.39.0
fqdn==1.5.1
fsspec @ file:///croot/fsspec_1714461537038/work
ftfy==6.2.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
greenlet==3.0.3
grpcio==1.51.3
humanfriendly==10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imagecodecs @ file:///croot/imagecodecs_1695064943445/work
imageio @ file:///croot/imageio_1707247282708/work
imageio-ffmpeg==0.5.1
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.11
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server==2.4.0
jupyter-server-fileid==0.8.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.6.1
jupyter-ydoc==0.2.3
jupyterlab==3.6.1
jupyterlab-language-pack-zh-cn==3.6.post0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.20.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kornia==0.7.3
kornia-rs==0.1.5
lazy-loader @ file:///croot/lazy_loader_1718176737906/work
lightning-utilities==0.11.8
lit==15.0.7
llvmlite==0.41.1
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
lpips==0.1.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
more-itertools==10.3.0
mpmath==1.3.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx @ file:///croot/networkx_1690561992265/work
notebook==6.5.3
notebook-shim==0.2.2
numba==0.58.1
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
oauthlib==3.2.2
opencv-python==4.5.4.60
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
partd @ file:///croot/partd_1698702562572/work
pexpect==4.8.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1721059439630/work
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==24.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.14.0
pyOpenSSL @ file:///croot/pyopenssl_1708380408460/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyWavelets @ file:///croot/pywavelets_1670425177960/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==25.0.1
regex==2024.7.24
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
scikit-fmm==2024.5.29
scikit-image @ file:///croot/scikit-image_1682528309703/work
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
-e git+https://gitclone.com/github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588#egg=segment_anything
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.31
stack-data==0.6.2
supervisor==4.2.5
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.5.0
tifffile @ file:///croot/tifffile_1695107451082/work
tinycss2==1.2.1
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
torch @ http://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp38-cp38-linux_x86_64.whl
torch-tools==0.1.5
torchmetrics==1.5.1
torchvision @ http://download.pytorch.org/whl/cu118/torchvision-0.15.1%2Bcu118-cp38-cp38-linux_x86_64.whl
tornado==6.2
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.12.2
tzdata==2024.1
uri-template==1.2.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
visdom==0.2.4
vtracer==0.6.11
wcwidth==0.2.13
webcolors==1.12
webencodings==0.5.1
webp==0.4.0
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
y-py==0.5.9
ypy-websocket==0.8.2
zipp @ file:///croot/zipp_1704206909481/work
